Initial commit to validate AppSync cache misses
Running the test in /test/api.test.ts authenticates the caller as the Test User in the Cognito User Pool. Their identity token is used as the authorization header for the Lambda authroizer. The test makes two calls against the AppSync API and both calls create a log trail when only the first one should (as the alls use the same API and token